pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ee-mcp"
version = "0.1.0"
description = "ExpressionEngine MCP Server"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Tyler Martin", email = "tylermartin2@example.com"},
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "python-dotenv>=1.0.0",
    "colorlog>=6.7.0",
    "httpx>=0.24.0",
]

[project.scripts]
ee-mcp = "ee_mcp.main:main"

[tool.setuptools]
packages = ["ee_mcp"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88
